refactor: profile.ts 의 경로를 바꾸고 그에 따른 다른 경로들 변경 (#101)
