+++ src/app/(profile)/profile/page.tsx
@@ -10,9 +10,9 @@ import MyDonationList from "@/components/profile/MyDonationList";
 import { useRouter } from "next/navigation";
 import { userStore } from "@/store/userStore";
 import { useMutation } from "@tanstack/react-query";
-import { patchUserProfile, uploadProfileImage } from "@/lib/profile";
 import EditProfileModal from "@/components/profile/EditProfileModal";
 import { debounce } from "@/lib/debounce";
+import { patchUserProfile, uploadProfileImage } from "@/lib/api/profile";
 
 export default function DonatorProfilePage() {
   const userData = userStore((state) => state.userData);
